+++ examples/graph_with_branching.py
@@ -34,7 +34,7 @@ async def end_task():
 
 if __name__ == "__main__":
     etl_graph = AsyncGraph()
-    etl_graph.add_node(extract, name="extract", max_tasks=1, queue_size=1_000)
+    etl_graph.add_node(extract, name="extract", max_tasks=1)
     etl_graph.add_node(transform, max_tasks=2)
     etl_graph.add_node(load, max_tasks=2)
     etl_graph.add_node(print_data, max_tasks=2)