+++ examples/halt_on_mismatched_num_of_args.py
@@ -25,7 +25,7 @@ async def load(data):
 
 if __name__ == "__main__":
     etl_graph = AsyncGraph()
-    etl_graph.add_node(extract, name="extract", max_tasks=1, queue_size=1_000)
+    etl_graph.add_node(extract, name="extract", max_tasks=1)
     etl_graph.add_node(transform)
     etl_graph.add_node(load)
     etl_graph.add_edge("extract", "transform")